refactor: onMount ロジックを app-state.svelte.ts の initApp() に抽出
- App.svelte の onMount 全ロジック（約400行）を initApp() に移動
- App.svelte の onMount は initApp({ deps }) の1行呼び出しに
- InitAppDeps interface で循環依存を回避（pullFromGitHub 等は deps 経由）
- atGuardEntry を appState に移動（updateUrlFromState と initApp 双方からアクセス）
- pushExitGuard() を initApp 内ローカル関数に移動
- $_() の3箇所を get(_) パターンに変更（.svelte.ts 内では $ prefix 不可）
- cleanup 関数の返却パターン維持
